requirements.txt
# LLM
openai>=1.30.0,<2.0.0

# CLI
typer>=0.12.0,<1.0.0
rich>=13.7.0,<14.0.0

# Config
python-dotenv>=1.0.0,<2.0.0

# Vector store (Part 2)
chromadb>=0.5.0,<1.0.0

# Embeddings (Part 2) — uses openai text-embedding-3-small
# Already covered by openai package above

# PDF parsing (Part 2)
pdfplumber>=0.11.0,<1.0.0

# Web framework (Part 3 / API)
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.29.0,<1.0.0

# Data validation
pydantic>=2.7.0,<3.0.0

# Testing
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0